requirements.txt
openai
langchain
transformers
torch
pandas
numpy
scikit-learn
beautifulsoup4
requests
flask
streamlit
plotly
dash
pytest
tqdm
sec-edgar
psycopg2
sentence_transformers
faiss-cpu
googlesearch-python
pdfplumber